Moved policy_for_dates, to the regular modules.
