+++ gettsim/tests/test_bp.py
@@ -5,7 +5,7 @@ from pandas.testing import assert_frame_equal
 
 from gettsim.benefits.benefit_checks import benefit_priority
 from gettsim.config import ROOT_DIR
-from gettsim.tests.policy_for_date import get_policies_for_date
+from gettsim.policy_for_date import get_policies_for_date
 
 
 INPUT_COLS = [
