+++ gettsim/tests/test_pensions.py
@@ -7,8 +7,8 @@ from gettsim.pensions import _rentenwert_from_2018
 from gettsim.pensions import _rentenwert_until_2017
 from gettsim.pensions import pensions
 from gettsim.pensions import update_earnings_points
+from gettsim.policy_for_date import get_policies_for_date
 from gettsim.tax_transfer import _apply_tax_transfer_func
-from gettsim.tests.policy_for_date import get_policies_for_date
 
 
 INPUT_COLS = [
