+++ gettsim/tests/test_tax_sched.py
@@ -4,9 +4,9 @@ import pytest
 from pandas.testing import assert_frame_equal
 
 from gettsim.config import ROOT_DIR
+from gettsim.policy_for_date import get_policies_for_date
 from gettsim.taxes.calc_taxes import tarif
 from gettsim.taxes.calc_taxes import tax_sched
-from gettsim.tests.policy_for_date import get_policies_for_date
 
 INPUT_COLS = [
     "pid",
